+++ clients/python/test/e2e/ci/bash/install_osparc_python_client.bash
