+++ clients/python/test/e2e/ci/bash/preprocess_client_config.bash
