+++ clients/python/test/e2e/ci/e2e/README.md
@@ -1,0 +1,1 @@
+Cli for running e2e tests
