+++ clients/python/test/e2e/ci/e2e/e2e/_utils.py
@@ -2,10 +2,12 @@ from enum import IntEnum
 from pathlib import Path
 
 import pytest
+import typer
+from pydantic import ValidationError
 
 # Paths -------------------------------------------------------
 
-_E2E_DIR: Path = Path(__file__).parent.parent.resolve()
+_E2E_DIR: Path = Path(__file__).parent.parent.parent.parent.resolve()
 _PYTHON_DIR: Path = _E2E_DIR.parent.parent
 _CI_DIR: Path = (_E2E_DIR / "ci").resolve()
 _PYTEST_INI: Path = (_E2E_DIR / "pytest.ini").resolve()
@@ -46,3 +48,14 @@ assert (
     )
     == set()
 )
+
+
+def handle_validation_error(func):
+    def wrapper(*args, **kwargs):
+        try:
+            return func(*args, **kwargs)
+        except ValidationError as e:
+            typer.echo(f"{e}", err=True)
+            raise typer.Exit(code=E2eExitCodes.CI_SCRIPT_FAILURE)
+
+    return wrapper
