+++ clients/python/test/e2e/ci/e2e/e2e/cli.py
@@ -1,0 +1,7 @@
+from typer import Typer
+
+from . import postprocess, preprocess
+
+cli = Typer()
+cli.add_typer(preprocess.cli, name="preprocess")
+cli.add_typer(postprocess.cli, name="postprocess")
