+++ clients/python/test/e2e/ci/e2e/pyproject.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "e2e"
+version = "0.1.0"
+description = ""
+authors = ["Mads Bisgaard <bisgaard@itis.swiss>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+typer = {extras = ["all"], version = "^0.9.0"}
+python = "^3.9"
+pydantic = "^2.5"
+pandas = "^2.1"
+pytest = "^7.4"
+
+[tool.poetry.scripts]
+e2e = "e2e.cli:cli"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
